fastlane_bot/tools/resources/NBTest/requirements.txt
psutil~=5.9.6
packaging==21.3
requests~=2.31.0
python-dateutil~=2.8.2
typing-extensions~=4.7.1
python-dotenv~=0.16.0
joblib~=1.2.0
pandas~=1.5.2
alchemy-sdk~=0.1.1
pyarrow~=11.0.0
networkx~=3.0
cvxpy~=1.3.1
matplotlib~=3.7.1
dataclass_wizard~=0.22.2
hexbytes~=0.3.1
click~=8.1.3
setuptools~=67.6.1
protobuf~=4.24.4
tqdm~=4.64.1
web3~=6.11.2
nest-asyncio~=1.5.8
